requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cors==1.0.1
cryptography==44.0.1
cssselect==1.3.0
cssselect2==0.7.0
cssutils==2.11.1
cycler==0.12.1
Django==5.1.5
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-sslserver==0.22
djangorestframework==3.15.2
filelock==3.17.0
fonttools==4.55.7
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
jwcrypto==1.5.6
jwt==1.3.1
kiwisolver==1.4.8
matplotlib==3.10.0
more-itertools==10.6.0
msal==1.32.0
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
premailer==3.10.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.1
pypdfium2==4.30.1
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
zope.event==5.0
zope.interface==7.2
zopfli==0.2.3.post1
